Extract SaaS provisioner service layer and fix the admin/instances route boundary
Move secret derivation, Helm args assembly, OpenRouter key lifecycle,
instance Secret management, deployment readiness, and the core
provision/start/stop/restart/uninstall/sync operations from
routes/provisioner.py into backend/services/provisioner_service.py.

routes/provisioner.py shrinks to auth + delegation. admin.py and
instances.py call the service directly instead of invoking another
route's handler with a synthesized 'Bearer PROVISIONER_API_KEY' header.

This also fixes a latent admin bug: admin_delete_account_complete called
uninstall_instance(instance_id=..., api_key=...), a signature that never
existed, so tenant deprovisioning during account deletion always raised
TypeError and was swallowed by the surrounding except.
