+++ saas-platform/platform-backend/src/backend/services/__init__.py
@@ -1,0 +1,1 @@
+"""Service layer shared by route modules."""
